fixed returned values and query type
