+++ packages/pytest-simcore/src/pytest_simcore/services_api_mocks_for_aiohttp_clients.py
@@ -120,7 +120,7 @@ def get_computation_cb(url, **kwargs) -> CallbackResult:
     node_states = FULL_PROJECT_NODE_STATES
 
     return CallbackResult(
-        status=202,
+        status=200,
         payload={
             "id": Path(url.path).name,
             "state": state,
